moved main functions into main.py file
